splitting apart ClusterControls, and incorporating feedback from bistline
