+++ app/javascript/components/visualization/controls/CreateAnnotation.js
