+++ app/lib/cluster_viz_service.rb
@@ -44,12 +44,18 @@ class ClusterVizService
   end
 
   # helper method to load spatial coordinate group names
+  # return an array of hashes with spatial file names and associated_clusters  specified, where associated_clusters is the names of
+  # the clusters
   def self.load_spatial_options(study)
+    # grab all the spatial files for this study, and create a name=>associations hash
     spatial_file_info = StudyFile.where(study: study, is_spatial: true, file_type: 'Cluster')
                                  .pluck(:name, :spatial_cluster_associations).to_h
+    # now grab any non spatial cluster files for the study that had ids specified in the spatial_cluster_associations above
+    # and put them in an id=>name hash
     associated_clusters = StudyFile.where(study: study, :id.in => spatial_file_info.map{ |si| si[1] }.flatten.uniq)
                                    .pluck(:id, :name)
                                    .map{ |a| [a[0].to_s, a[1]] }.to_h
+    # now return an array of objects with names and associated cluster names
     spatial_file_info.map do |cluster|
       associated_cluster_names = cluster[1].map{ |id| associated_clusters[id] }
       { name: cluster[0], associated_clusters: associated_cluster_names }
